% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glabraster.antarctica.R
\docType{data}
\name{Glabraster.antarctica}
\alias{Glabraster.antarctica}
\title{Presence-only records of the sea star \emph{Glabraster antarctica} (Southern Ocean)}
\format{
A two columns table (longitude, latitude)
}
\source{
Moreau, C., Mah, C., Agüera, A., Améziane, N., Barnes, D., Crokaert, G., ... & Jażdżewska, A. (2018). Antarctic and sub-Antarctic Asteroidea database. ZooKeys, (747), 141.
}
\usage{
data(Glabraster.antarctica)
}
\description{
Dataset that contains the presence data of the sea star species \emph{Glabraster antarctica} reported in the Southern Ocean. The detailed description of the dataset is available in Moreau et al. (2018)
}
\examples{
library(SDMPlay)
data(Glabraster.antarctica)
head(Glabraster.antarctica)

}
\keyword{datasets}
