% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Odontaster.validus.R
\docType{data}
\name{Odontaster.validus}
\alias{Odontaster.validus}
\title{Presence-only records of the sea star \emph{Odontaster validus} (Southern Ocean)}
\format{
A two columns table (longitude, latitude)
}
\source{
Moreau, C., Mah, C., Agüera, A., Améziane, N., Barnes, D., Crokaert, G., ... & Jażdżewska, A. (2018). Antarctic and sub-Antarctic Asteroidea database. ZooKeys, (747), 141.
}
\usage{
data(Odontaster.validus)
}
\description{
Dataset that contains the presence data of the sea star species \emph{Odontaster validus} reported in the Southern Ocean. The detailed description of the dataset is available in Moreau et al. (2018)
}
\examples{
data(Odontaster.validus)
head(Odontaster.validus)

}
\keyword{datasets}
