% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap.R
\docType{data}
\name{worldmap}
\alias{worldmap}
\title{Worldmap}
\format{
csv file
}
\source{
[ADD website](https://data.aad.gov.au/metadata/records/fulldisplay/environmental_layers)
}
\usage{
data("worldmap")
}
\description{
csv file to draw worldmap on maps
}
\examples{
data("worldmap")
}
\keyword{datasets}
